These scripts are identical to r7761 on the tiegcm branch. There is a subversion problem that will not let me merge the branch back onto the trunk, so the log files for the revisions on the branch are only available on the branch.
In summary - these scripts enforce some assumptions about filenames
and such through the use of 'sed' on input.nml and tiegcm.nml.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7768 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: c1ff0492dc6fdfb2f7a15053cf7f77c0f64f0e3c

+++ models/tiegcm/shell_scripts/run_filter_async4.csh
@@ -6,23 +6,23 @@
 #
 # DART $Id$
 #
-#=============================================================================
-# This block of directives constitutes the preamble for the LSF queuing system
-# LSF is used on the IBM   Linux cluster 'lightning'
-# LSF is used on the IMAGe Linux cluster 'coral'
-# LSF is used on the IBM   'bluevista'
-# The queues on lightning and bluevista are supposed to be similar.
-#
-# the normal way to submit to the queue is:    bsub < run_filter.csh
-#
-# an explanation of the most common directives follows:
-# -J Job name (master script job.csh presumes filter_server.xxxx.log)
-# -o STDOUT filename
-# -e STDERR filename
-# -P      account
-# -q queue    cheapest == [standby, economy, (regular,debug), premium] == $$$$
-# -n number of processors  (really)
-##=============================================================================
+##------------------------------------------------------------------------------
+## This block of directives constitutes the preamble for the LSF queuing system
+## LSF is used on the IBM   Linux cluster 'lightning'
+## LSF is used on the IMAGe Linux cluster 'coral'
+## LSF is used on the IBM   'bluevista'
+## The queues on lightning and bluevista are supposed to be similar.
+##
+## the normal way to submit to the queue is:    bsub < run_filter.csh
+##
+## an explanation of the most common directives follows:
+## -J Job name (master script job.csh presumes filter_server.xxxx.log)
+## -o STDOUT filename
+## -e STDERR filename
+## -P      account
+## -q queue    cheapest == [standby, economy, (regular,debug), premium] == $$$$
+## -n number of processors  (really)
+##
 #BSUB -J filter
 #BSUB -o filter.%J.log
 #BSUB -P P3507xxxx
@@ -31,7 +31,7 @@
 #BSUB -W 0:30
 #BSUB -N -u ${USER}@ucar.edu
 #
-##=============================================================================
+##------------------------------------------------------------------------------
 ## This block of directives constitutes the preamble for the PBS queuing system
 ## PBS is used on the CGD Linux cluster 'bangkok'
 ## PBS is used on the CGD Linux cluster 'calgary'
@@ -48,7 +48,7 @@
 ##                     and calgary, there is no way to 'share' the processors
 ##                     on the node with another job, so you might as well use
 ##                     them both. (ppn == Processors Per Node)
-##=============================================================================
+##
 #PBS -N filter
 #PBS -r n
 #PBS -e filter.err
@@ -56,19 +56,7 @@
 #PBS -q dedicated
 #PBS -l nodes=10:ppn=2
 
-# if async=2, e.g. you are going to run './modelxxx', single process
-# (or possibly 'mpirun -np 1 ./modelxxx'), so each processor advances
-# one ensemble independently of the others, leave this as false.
-#
-# if async=4, e.g. all the processors advance each modelxxx in turn with
-# mpirun -np 64 modelxxx (or whatever) for as many ensembles as you have,
-# set this to "true"
-
-# if async=4, also check that the call to advance_model.csh
-# has the right number of ensemble members below; it must match
-# the input.nml number.
-
-set parallel_model = "true"
+#===============================================================================
 
 # Determine the number of ensemble members from input.nml,
 # it may exist in more than one place.
@@ -84,6 +72,17 @@ endif
 set ENSEMBLESTRING = `grep -A 42 filter_nml input.nml | grep ens_size`
 set NUM_ENS = `echo $ENSEMBLESTRING[3] | sed -e "s#,##"`
 
+# FIXME ... read the async value from input.nml and set parallel_model accordingly.
+# if async=2, e.g. you are going to run './modelxxx', single process
+# (or possibly 'mpirun -np 1 ./modelxxx'), so each processor advances
+# one ensemble independently of the others, leave this as false.
+#
+# if async=4, e.g. all the processors advance each modelxxx in turn with
+# mpirun -np 64 modelxxx (or whatever) for as many ensembles as you have,
+# set this to "true"
+
+set parallel_model = "true"
+
 # A common strategy for the beginning is to check for the existence of
 # some variables that get set by the different queuing mechanisms.
 # This way, we know which queuing mechanism we are working with,
@@ -95,135 +94,14 @@ if ($?LS_SUBCWD) then
     # LSF has a list of processors already in a variable (LSB_HOSTS)
     # alias submit 'bsub < \!*'
     echo "LSF - using mpirun.lsf for execution"
-
-    # each filter task advances the ensembles, each running on 1 proc.
-    if ( "$parallel_model" == "false" ) then
-
-       mpirun.lsf ./filter
-
-    else
-
-    # filter runs in parallel until time to do a model advance,
-    # and then this script starts up the modelxxx jobs, each one
-    # running in parallel. then it runs wakeup_filter to wake
-    # up filter so it can continue.
-
-      \rm -f model_to_filter.lock filter_to_model.lock
-      mkfifo model_to_filter.lock filter_to_model.lock
-
-      set filterhome = ~/.filter$$
-      if ( ! -e $filterhome) mkdir $filterhome
-
-      # this starts filter but also returns control back to
-      # this script immediately.
-
-      (setenv HOME $filterhome; mpirun.lsf ./filter) &
-
-      while ( -e filter_to_model.lock )
-
-        set todo=`cat < filter_to_model.lock`
-        echo "todo received, value = ${todo}"
-
-        if ( "${todo}" == "finished" ) then
-          echo "main script: filter done."
-          wait
-          break
-
-        else if ( "${todo}" == "advance" ) then
-
-          # the second number below must match the number
-          # of ensembles. Also, in input.nml, the advance model
-          # command must have -np N with N equal to the number
-          # of processors this job is using.
-
-          echo "calling model advance now:"
-          ./advance_model.csh 0 ${NUM_ENS} filter_control00000 || exit 9
-
-          echo "restarting filter."
-          mpirun.lsf ./wakeup_filter
-
-        else
-
-          echo "main script: unexpected value received."
-          break
-
-        endif
-
-      end
-
-      echo "filter finished, removing pipes."
-      \rm -f model_to_filter.lock filter_to_model.lock
-
-      if ( -d $filterhome) rmdir $filterhome
-    endif
-
+    setenv MPICMD mpirun.lsf
 
 else if ($?PBS_O_WORKDIR) then
 
     # PBS has a list of processors in a file whose name is (PBS_NODEFILE)
     # alias submit 'qsub \!*'
     echo "PBS - using mpirun for execution"
-
-    # each filter task advances the ensembles, each running on 1 proc.
-    if ( "$parallel_model" == "false" ) then
-
-      mpirun ./filter
-
-    else
-
-    # filter runs in parallel until time to do a model advance,
-    # and then this script starts up the modelxxx jobs, each one
-    # running in parallel. then it runs wakeup_filter to wake
-    # up filter so it can continue.
-
-      \rm -f model_to_filter.lock filter_to_model.lock
-      mkfifo model_to_filter.lock filter_to_model.lock
-
-      set filterhome = ~/.filter
-      if ( ! -e $filterhome) mkdir $filterhome
-
-      # this starts filter but also returns control back to
-      # this script immediately.
-
-      (setenv HOME $filterhome; mpirun ./filter) &
-
-      while ( -e filter_to_model.lock )
-
-        set todo=`cat < filter_to_model.lock`
-        echo "todo received, value = ${todo}"
-
-        if ( "${todo}" == "finished" ) then
-          echo "main script: filter done."
-          wait
-          break
-
-        else if ( "${todo}" == "advance" ) then
-
-          # the second number below must match the number
-          # of ensembles. Also, in input.nml, the advance model
-          # command must have -np N with N equal to the number
-          # of processors this job is using.
-
-          echo "calling model advance now:"
-          ./advance_model.csh 0 ${NUM_ENS} filter_control00000 || exit 9
-
-          echo "restarting filter."
-          mpirun ./wakeup_filter
-
-        else
-
-          echo "main script: unexpected value received."
-          break
-
-        endif
-
-      end
-
-      echo "filter finished, removing pipes."
-      \rm -f model_to_filter.lock filter_to_model.lock
-
-      if ( -d $filterhome) rmdir $filterhome
-    endif
+    setenv MPICMD mpirun
 
 else
 
@@ -254,58 +132,68 @@ else
 
     echo "MPICMD = ${MPICMD}"
 
-    # filter runs in parallel until time to do a model advance,
-    # and then this script starts up the modelxxx jobs, each one
-    # running in parallel. then it runs wakeup_filter to wake
-    # up filter so it can continue.
+endif
 
-    \rm -f model_to_filter.lock filter_to_model.lock
-    mkfifo model_to_filter.lock filter_to_model.lock
+#-------------------------------------------------------------------------------
+# Everything below this separator should not need to be modified if everything
+# above the separator is set correctly.
+#-------------------------------------------------------------------------------
 
-    set filterhome = ~/.filter$$
-    if ( ! -e $filterhome) mkdir $filterhome
+if ( "$parallel_model" == "false" ) then
 
-    # this starts filter but also returns control back to
-    # this script immediately.
+   # each filter task advances the ensembles, each running on 1 proc.
 
-    (setenv HOME $filterhome; ${MPICMD} ./filter) &
+   ${MPICMD} ./filter
 
-    while ( -e filter_to_model.lock )
+else
 
-        set todo=`cat < filter_to_model.lock`
-        echo "todo received, value = ${todo}"
+   # filter runs in parallel until time to do a model advance,
+   # and then this script starts up the modelxxx jobs, each one
+   # running in parallel. then it runs wakeup_filter to wake
+   # up filter so it can continue. The communication happens through
+   # 'named pipes' created by the mkfifo command.
 
-        if ( "${todo}" == "finished" ) then
-          echo "main script: filter done."
-          wait
-          break
+   \rm -f model_to_filter.lock filter_to_model.lock
+   mkfifo model_to_filter.lock filter_to_model.lock
 
-        else if ( "${todo}" == "advance" ) then
+   set filterhome = ~/.filter$$
+   if ( ! -e $filterhome) mkdir $filterhome
 
-          # the second number below must match the number
-          # of ensembles. Also, in input.nml, the advance model
-          # command must have -np N with N equal to the number
-          # of processors this job is using.
+   # start filter and immediately return control back to this script
 
-          echo "calling model advance now:"
-          ./advance_model.csh 0 ${NUM_ENS} filter_control00000 || exit 9
+   (setenv HOME $filterhome; ${MPICMD} ./filter) &
 
-          echo "restarting filter."
-          ${MPICMD} ./wakeup_filter
+   while ( -e filter_to_model.lock )
 
-        else
+      set todo=`cat < filter_to_model.lock`
+      echo "todo received, value = ${todo}"
 
+      if ( "${todo}" == "finished" ) then
+         echo "main script: filter done."
+         wait
+         break
+
+      else if ( "${todo}" == "advance" ) then
+
+         # FIXME : in input.nml, the advance model command must
+         # have -np N with N equal to the number of processors this job is using.
+
+         echo "calling model advance now:"
+         ./advance_model.csh 0 ${NUM_ENS} filter_control00000 || exit 9
+
+         echo "restarting filter."
+         ${MPICMD} ./wakeup_filter
+
+      else
           echo "main script: unexpected value received."
           break
+      endif
+   end
 
-        endif
+   echo "filter finished, removing pipes."
+   \rm -f model_to_filter.lock filter_to_model.lock
 
-    end
-
-    echo "filter finished, removing pipes."
-    \rm -f model_to_filter.lock filter_to_model.lock
-
-    if ( -d $filterhome) rmdir $filterhome
+   if ( -d $filterhome) \rmdir $filterhome
 
 endif
